+++ groovy/src/test/groovy/cucumber/runtime/groovy/compiled_stepdefs.groovy
@@ -13,7 +13,7 @@ Given(~'^I have (\\d+) apples in my belly') { int apples ->
     lastAte('apples')
 }
 
-Given(~'^a big basket with cukes') {->
+Given(~'^a big basket with cukes') { ->
 }
 
 Given(~'^the following table:$') { table ->