+++ groovy/src/test/java/cucumber/runtime/groovy/RunCukesTest.java
@@ -4,5 +4,6 @@ import cucumber.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
+@Cucumber.Options(tags="@focus")
 public class RunCukesTest {
 }